TruthLens-AI/backend_server/requirements.txt
faiss-cpu
dotenv
ddgs
sentence-transformers
huggingface_hub
fastapi
uvicorn
pydantic